# docs/07_interfaces/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevPilot Local API Contract v1",
    "version": "1.0.0-implemented-initial",
    "description": "Static OpenAPI contract synchronized with the implemented-initial local FastAPI MVP. FUNC-SPRINT-67 implements the server adapter on localhost; Sprint 68 will add token/CORS hardening."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Future local-only server. Not active in Sprint 66."
    }
  ],
...
  "paths": {
...
    "/api/v1/review/code": {
      "post": {
        "operationId": "review_code",
        "summary": "API-REVIEW-CODE: review.code",
        "description": "Contract-only endpoint mapped to `ReviewApplicationService.code_review`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "review"
        ],
        "x-devpilot-api-id": "API-REVIEW-CODE",
        "x-devpilot-operation": "review.code",
        "x-devpilot-domain-service": "ReviewApplicationService.code_review",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "token-planned-sprint-68",
        "x-devpilot-policy": "CODE_REVIEW_DRY_RUN_ALLOW",
        "x-devpilot-command-equivalent": "python -m devpilot_core code-review . --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplicationRequest"
                  }
                ],
                "properties": {
                  "operation": {
                    "const": "review.code"
                  }
                }
              },
              "example": {
                "operation": "review.code",
                "payload": {},
                "client": "api-local",
                "dry_run": true
              }
            }
          }
        }
      }
    },
    "/api/v1/refactor/plan": {
      "post": {
        "operationId": "refactor_plan",
        "summary": "API-REFACTOR-PLAN: refactor.plan",
        "description": "Contract-only endpoint mapped to `RefactorApplicationService.plan`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "refactor"
        ],
        "x-devpilot-api-id": "API-REFACTOR-PLAN",
        "x-devpilot-operation": "refactor.plan",
        "x-devpilot-domain-service": "RefactorApplicationService.plan",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "token-planned-sprint-68",
        "x-devpilot-policy": "plan-only; no patch execution",
        "x-devpilot-command-equivalent": "python -m devpilot_core refactor-plan . --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplicationRequest"
                  }
                ],
                "properties": {
                  "operation": {
                    "const": "refactor.plan"
                  }
                }
              },
              "example": {
                "operation": "refactor.plan",
                "payload": {},
                "client": "api-local",
                "dry_run": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Finding": {
        "type": "object",
        "required": [
          "id",
          "message",
          "severity"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "fail",
              "block",
              "error"
            ]
          },
          "path": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "ApplicationRequest": {
        "type": "object",
        "required": [
          "operation",
          "payload",
          "client",
          "dry_run"
        ],
        "properties": {
          "operation": {
            "type": "string"
          },
          "payload": {
            "type": "object",
            "additionalProperties": true
          },
          "client": {
            "type": "string",
            "default": "api-local"
          },
          "dry_run": {
            "type": "boolean",
            "default": true
          }
        },
        "additionalProperties": true
      },
      "ApplicationResponse": {
        "type": "object",
        "required": [
          "contract",
          "schema_version",
          "operation",
          "ok",
          "exit_code",
          "message",
          "data",
          "findings",
          "generated_at"
        ],
        "properties": {
          "contract": {
            "type": "string",
            "const": "DevPilotApplicationResponse"
          },
          "schema_version": {
            "type": "string"
          },
          "operation": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "exit_code": {
            "type": "integer",
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Finding"
            }
          },
          "generated_at": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "ErrorApplicationResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApplicationResponse"
          }
        ],
        "description": "Error envelope preserving ApplicationResponse semantics. Non-2xx responses must still carry operation, exit_code and findings."
      }
    },
    "securitySchemes": {
      "LocalTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-DevPilot-Local-Token",
        "description": "Planned for FUNC-SPRINT-68. Not enforced by Sprint 66 because no HTTP server exists."
      }
    }
  }
}